regular/2025-03/docker/code/toDoList/app/requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
Werkzeug==2.2.2